requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<3.13
psycopg2>=2.9